+++ rss-puzzle/src/components/main/imageNames.ts
@@ -1,47 +1,47 @@
 export const imageNames1 = [
-    '9th_wave',
-    'abbati2',
-    'arabs',
-    'campo',
-    'citywall',
-    'coastal',
-    'deerhunt',
-    'deerlake',
-    'distress',
-    'edgewood',
-    'extensiv',
-    'extensiv_1',
-    'firework',
-    'fishing',
-    'giuseppe',
-    'ice_land',
-    'italiana',
-    'italianb',
-    'kilarney',
-    'landmose',
-    'landsca3',
-    'landscap',
-    'landscap_1',
-    'landscap_2',
-    'railway',
-    'rateship',
-    'river_la',
-    'riverla2',
-    'rome',
-    'rome1',
-    'scene',
-    'shipcalm',
-    'shipping',
-    'shipping_1',
-    'skating',
-    'tivoli',
-    'vessels1',
-    'view_stp',
-    'viewvien',
-    'viewvlaa',
-    'waterfal',
-    'winter_l',
-    'winterla',
-    'winterla_1',
-    'woodedla',
-  ];
+  '9th_wave',
+  'abbati2',
+  'arabs',
+  'campo',
+  'citywall',
+  'coastal',
+  'deerhunt',
+  'deerlake',
+  'distress',
+  'edgewood',
+  'extensiv',
+  'extensiv_1',
+  'firework',
+  'fishing',
+  'giuseppe',
+  'ice_land',
+  'italiana',
+  'italianb',
+  'kilarney',
+  'landmose',
+  'landsca3',
+  'landscap',
+  'landscap_1',
+  'landscap_2',
+  'railway',
+  'rateship',
+  'river_la',
+  'riverla2',
+  'rome',
+  'rome1',
+  'scene',
+  'shipcalm',
+  'shipping',
+  'shipping_1',
+  'skating',
+  'tivoli',
+  'vessels1',
+  'view_stp',
+  'viewvien',
+  'viewvlaa',
+  'waterfal',
+  'winter_l',
+  'winterla',
+  'winterla_1',
+  'woodedla',
+];
