+++ rss-puzzle/src/components/main/interfaces.ts
@@ -1,0 +1,12 @@
+export interface IWord {
+ textExample: string;
+ textExampleTranslate: string;
+}
+
+export interface IRound {
+ words: IWord[];
+}
+
+export interface IWordData {
+ rounds: IRound[];
+}
